fix: split up city wallet and token mgmt in auth tests
